+++ config/config.exs
@@ -5,6 +5,9 @@ config :torngen,
   out_dir: "lib/torngen_client/",
   generator: :elixir
 
+config :tornex,
+  client: Tornex.HTTP.FinchClient
+
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
 import_config "#{config_env()}.exs"