+++ test/test_helper.exs
@@ -12,4 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+Supervisor.start_link([Tornex.HTTP.FinchClient], strategy: :one_for_one)
+
 ExUnit.start()